fix: resolve SonarCloud issues in FileController, HealthCheck service, and bean naming

+++ src/main/java/com/loadbalancer/service/HealthCheckService.java
@@ -89,17 +89,21 @@ public class HealthCheckService {
 
     try {
       restTemplate.getForEntity(url, String.class);
-      double duration = (double)(System.currentTimeMillis() - startTime);
+      // No cast needed as System.currentTimeMillis() returns a long, and subtraction of longs is a long
+      // which can be implicitly converted to double for the recordRequest method
+      double duration = System.currentTimeMillis() - startTime;
       metricsService.recordRequest(node.getContainerId().toString(), true, duration);
     } catch (Exception e) {
       isHealthy = false;
       statusMessage = e.getMessage();
-      double duration = (double)(System.currentTimeMillis() - startTime);
+      // No cast needed here either
+      double duration = System.currentTimeMillis() - startTime;
       metricsService.recordRequest(node.getContainerId().toString(), false, duration);
     }
 
-    // Calculate response time as a double
-    double responseTime = (double)(System.currentTimeMillis() - startTime);
+    // No cast needed here either - the subtraction of two longs results in a long
+    // which can be implicitly converted to double
+    double responseTime = System.currentTimeMillis() - startTime;
 
     return HealthStatus.builder()
             .nodeId(node.getContainerId().toString())
